Make theories the last thing to be compiled.
This should make life easier for people writing new theories in the
core.

Also introduced a "dummy theory" that just complains when you ask it to
do anything - this is the theory that is selected when you start the
core.  So to do anything, you need to change the theory.