Remove priority code from the GUI.
It breaks things, especially normalisation.  This needs to be fixed so
that rule priorities are cached, rather than being fetched from the core
over and over again (in the rendering code!!!) with things from
different threads interfering with each other.

Normalisation is quite brittle, as it is multi-threaded, and the Core
class is NOT threadsafe.