Matching: deal with match conflicts at the end of the wire
In the case where we attempted to match a wire against a wire in the
target graph whose endpoint had already been matched (eg: Z-X-Z against
X=Z), and these matchings did not agree, VInjEndo would throw an
exception as we tried to break injectivity.

In the algorithm description, this check was not made clear, as it was
implied by the phrase "valid wire matchings".