Fix renaming-apart a rule before application
Issue: suppose we have a rule with boundary {a,b} and a graph g such
that, when the graph is normalised, it introduces a vertex b, and a
matches this new b.  Then doing the graph substitution would involve the
renames a->b and b->c (where b matches a vertex c).  This is something
CRnm can't cope with.  So we have to normalise g and only _then_
rename-apart the rule.

regression-tests.ML contains a test case for this.