Put the right graphs in tgt and full_tgt in match state
full_tgt was getting the graph that should have been in tgt and vice
versa. This meant that matching would always be done against the full
graph, and exceptions could be raised in graph_subst when trying to
remove the matched vertices from full_tgt (which might not contain all
of the image of the match).