Use a set, not a subgraph, to control where the inner match search looks
Rather than maintaining both a tgt and a full_tgt graph in the match
state, we only retain a tgt (which is what full_tgt used to be), and
also a set tgt_verts (aka T) of unmatched vertices to match onto.

When we have !-box matching, this will be used for more fine-grained
control of where the matching happens.