Fix LaTeX (TikZ) output generation
Vertex.getVertexType().toString() would not return the typename, but a
Java magic string, and Vertex.getVertexType() would be null for
wire/boundary vertices, leading to a null pointer exception.

Both these issues are now fixed, by checking for null (and using "wire"
for the type name in that case) and by implement toString() for
VertexType.

Fixes #111