MAJOR: Now can handle composite types
Rather than a node's data being a single type it is now a vector of available types.
E.g. setting the data to "angle_expr, boolean" will coerce the data into an angle and a boolean
Matching works _by type_, so `\alpha` in a position that is supposed to be an angle will match only other expressions that are also meant to be angles.
The theory "composite.qtheory" is given as an example